Use correct name to fix projects setup query (#181)
Fixes #180.

Previously, this was using `starterName` in the `name` field when
preparing the projects query. As a result, the project would not be
found, and it would fail.

Additionally, after failing, it would have already created the database,
meaning that when the script is re-run, `BYPASS_FINISH` would be set to
true since the file at the dbPath exists. This results in the initial
user not being setup correctly.

By fixing the query, the projects can be found, and the sqlite setup
completes correctly and sets `.finish_setup` for use in creation of the
first user as admin. (Confirmed working.)